test: fix integration tests
