+++ jest.integration.ts
@@ -4,6 +4,7 @@
  * @see https://jestjs.io/docs/en/configuration
  */
 module.exports = {
+  preset: 'ts-jest',
   rootDir: 'test',
   testEnvironment: 'node',
 };
