+++ package.json
@@ -19,7 +19,7 @@
     "release": "standard-version --no-verify",
     "test": "jest",
     "test:ci": "jest --coverage --ci",
-    "test:integration": "npm run build && jest --config=jest.integration.js",
+    "test:integration": "npm run build && jest --config=jest.integration.ts",
     "test:watch": "npm test -- --watch"
   },
   "repository": {
